UseEffect Debug
so setInterval calls itself so this a good reminder the callback inside a callback  inside the callback of a useEffect can be async and needs to be when fetching data. Setting storage can be done with useEffect and if there is no useEffect we use a different cookie as a starter. Sooooo basically alll sideEffects and stuff like that